The goal of this change is to make it so that the Future object won't get freed until the observe callbacks are finished executing.
This is useful when you have code as follows that gets a future within
the scope of a method, because when the scope of `someFunc()` ends
`future` is freed. This can cause race conditions with the `didSet`
property observer in the Future implementation.

```
func someFunc() {
  let future = somethingThatReturnsAFuture()
  future.observe(with: { result in
    // Do stuff
  }
}
```

This can of course be worked around in the client app by capturing a
strong reference to `future` inside of the observe callback as follows.

```
func someFunc() {
  let future = somethingThatReturnsAFuture()
  future.observe(with: { result in
    let _ = future // capture, creating strong reference
    // Do stuff
  }
}
```

Creating that strong reference in the scope of the observe callback
closure makes it so that `future` can't be freed at least until the
observe callback is complete.

I have taken a very similar approach inside of `observe` method's
implementation by creating a wrapping closure that simply captures a
strong reference to `self` and then executes the observe callback. Which
should make sure the future, in this case `self`, isn't freed at least
until all the observe callbacks have completed.

I think this is a better solution than forcing consuming application
developers to have to worry about managing the lifetime of the future
instance using hackery like above but all across their application.

I don't see any real downsides to doing this either. Though it is always
possible I missed something.

To be safe I also added a didSet weak self lambda wrapping the observer
callback to make sure that if self is gone when the didSet callback gets
executed it doesn't crash the app.

Beyond that I added a `cancel` method to explicitly be able to tell the
future you no longer care about triggering observers when the promise
finishes.