Change chained to require the same error type
I did this so that the error type of the future produced by the provided
closure would have to be the same as the error type of the parenting
future. I also ended up having to remove the do/catch handling of the
closure and force the closure to not throw. This in turn forces the
consumer to handle mapping all throws inside the closure to the error
type of the future returned from the closure.

I think this not only makes this chained method work correctly in terms
of responsibilities and concerns but also forces the responsibility of
handling and mapping the throws to error types to the correct location.

[changelog]
changed: Future.chained(closure:) to require the same error type as the parenting Future

ps-id: 3300FC52-D497-403A-AC9C-93BB7F48FB26