Fix memory leak where the Futures are not freed
After some testing the issue seems to be with the explicit retain cycle
that we setup in callbackWrappers. Specifically, it looks like the
explicit retain cycle wasn't being broken. To resolve this I modified
the report(result:) method to clear out all the callbacks (and their
callbackWrappers) once the callbacks completed. This in turn breaks the
explicit retain cycle and allows the Future to be deallocated.

[changelog]
fixed: memory leak where Futures were not being freed